Se modifico el prototipo de la funcion a un tipo de retorno tipo double.

+++ main.c
@@ -9,7 +9,7 @@ extern int pop(int *stack, int pos);
 extern int peek(int *stack, int pos);
 extern void print(int *stack, int pos);
 extern void _compara( char in[256]);
-extern float cos( float angle );
+extern double cos( float angle );
 
 int main(void){
   int *stack;
@@ -17,7 +17,7 @@ int main(void){
   int val = 0;
 
   float value = 0;
-  float res = 0;
+  double res = 0;
   printf("Stack.\n\n");
   stack = malloc( INIT * sizeof(char) );
   // while ( in  == 0 ){
